Переносит функционал взаимодействия с формой в точку входа
