+++ js/main.js
@@ -1,7 +1,6 @@
 import {createPicturesData} from './data.js';
 import {renderGallery} from './gallery/main.js';
-import {renderUpload} from './upload/main.js';
+import './upload/main.js';
 
 renderGallery(createPicturesData());
 
-renderUpload();
