+++ js/upload/popup.js
@@ -1,7 +1,3 @@
-import {validateForm, resetForm} from './validation.js';
-
-const pictureUpload = document.querySelector('.img-upload__input');
-const form = document.querySelector('.img-upload__form');
 const popup = document.querySelector('.img-upload__overlay');
 const closeButton = document.querySelector('.img-upload__cancel');
 
@@ -28,27 +24,10 @@ const openPopup = () => {
 const closePopup = () => {
   popup.classList.add('hidden');
   document.body.classList.remove('modal-open');
-  pictureUpload.value = '';
   document.removeEventListener('keydown', onDocumentKeydown);
   document.removeEventListener('click', onOverlayClick);
 };
 
 closeButton.addEventListener('click', () => closePopup());
 
-const renderPopup = () => {
-
-  pictureUpload.addEventListener('change', () => {
-    openPopup();
-  });
-
-  form.addEventListener('submit', (event) => {
-    event.preventDefault();
-    validateForm();
-  });
-
-  form.addEventListener('reset', () => {
-    resetForm();
-  });
-};
-
-export {renderPopup};
+export {openPopup};
